remove last? np.float refs

+++ imfun/track.py
@@ -274,8 +274,8 @@ class LCV_Contours:
             inside = (rc > l) * (rc < u)
             l,u = int(l),int(u)
             outside = ~inside # outside is a negation of inside
-            c1 = np.sum(self.U[inside,j])/np.float(np.sum(inside+1e-12))
-            c2 = np.sum(self.U[outside,j])/np.float(np.sum(outside+1e-12))
+            c1 = np.sum(self.U[inside,j])/np.sum(inside+1e-12)
+            c2 = np.sum(self.U[outside,j])/np.sum(outside+1e-12)
             v = np.abs(self.U[:,j]-c1) - np.abs(self.U[:,j]-c2)
             vmax = abs(v).max()
             a[j] += self.lam*v[l]*self.weights[j]/vmax
